+++ cmn-data-svc-server/db-scripts/cds-mysql-downgrade-1.16-to-1.15.sql
@@ -1,0 +1,36 @@
+-- ===============LICENSE_START=======================================================
+-- Acumos Apache-2.0
+-- ===================================================================================
+-- Copyright (C) 2017-2018 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+-- ===================================================================================
+-- This Acumos software file is distributed by AT&T and Tech Mahindra
+-- under the Apache License, Version 2.0 (the "License");
+-- you may not use this file except in compliance with the License.
+-- You may obtain a copy of the License at
+--
+--      http://www.apache.org/licenses/LICENSE-2.0
+--
+-- This file is distributed on an "AS IS" BASIS,
+-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+-- See the License for the specific language governing permissions and
+-- limitations under the License.
+-- ===============LICENSE_END=========================================================
+
+-- Script to downgrade database used by the Common Data Service
+-- FROM version 1.16.x TO version 1.15.x.
+-- No database is specified to allow flexible deployment!
+
+-- 7
+ALTER TABLE C_SOLUTION_REV DROP COLUMN PUBLISHER;
+-- 6
+ALTER TABLE C_SOLUTION ADD COLUMN PROVIDER CHAR(64);
+-- 5
+ALTER TABLE C_PEER_SUB CHANGE USER_ID OWNER_ID CHAR(36) not null;
+-- 4
+ALTER TABLE C_ARTIFACT CHANGE USER_ID OWNER_ID CHAR(36) not null;
+-- 3
+ALTER TABLE C_SOLUTION_REV CHANGE USER_ID OWNER_ID CHAR(36) not null;
+-- 2
+ALTER TABLE C_SOLUTION CHANGE USER_ID OWNER_ID CHAR(36) not null;
+-- 1
+ALTER TABLE C_SOLUTION DROP COLUMN PICTURE;