+++ cmn-data-svc-server/src/main/java/org/acumos/cds/controller/SolutionController.java
@@ -333,13 +333,15 @@ public class SolutionController extends AbstractController {
 			// All remaining parameters are optional
 			String[] nameKws = getOptStringArray(CCDSConstants.SEARCH_NAME, queryParameters);
 			String[] descKws = getOptStringArray(CCDSConstants.SEARCH_DESC, queryParameters);
-			String[] ownerIds = getOptStringArray(CCDSConstants.SEARCH_OWNERS, queryParameters);
+			String[] userIds = getOptStringArray(CCDSConstants.SEARCH_USERS, queryParameters);
 			String[] modelTypeCodes = getOptStringArray(CCDSConstants.SEARCH_MODEL_TYPES, queryParameters);
 			String[] accTypeCodes = getOptStringArray(CCDSConstants.SEARCH_ACCESS_TYPES, queryParameters);
 			String[] valStatusCodes = getOptStringArray(CCDSConstants.SEARCH_VAL_STATUSES, queryParameters);
 			String[] tags = getOptStringArray(CCDSConstants.SEARCH_TAGS, queryParameters);
-			Object result = solutionSearchService.findPortalSolutions(nameKws, descKws, active, ownerIds,
-					modelTypeCodes, accTypeCodes, valStatusCodes, tags, pageRequest);
+			String[] authKws = getOptStringArray(CCDSConstants.SEARCH_AUTH, queryParameters);
+			String[] pubKws = getOptStringArray(CCDSConstants.SEARCH_PUB, queryParameters);
+			Object result = solutionSearchService.findPortalSolutions(nameKws, descKws, active, userIds, modelTypeCodes,
+					accTypeCodes, valStatusCodes, tags, authKws, pubKws, pageRequest);
 			logger.audit(beginDate, "findPortalSolutions: query {}", queryParameters);
 			return result;
 		} catch (Exception ex) {
@@ -373,7 +375,7 @@ public class SolutionController extends AbstractController {
 		try {
 			// These parameters are required
 			String activeString = queryParameters.getFirst(CCDSConstants.SEARCH_ACTIVE);
-			String userId = queryParameters.getFirst(CCDSConstants.SEARCH_OWNERS);
+			String userId = queryParameters.getFirst(CCDSConstants.SEARCH_USERS);
 			if (activeString == null || activeString.length() == 0 || userId == null || userId.length() == 0)
 				throw new IllegalArgumentException("Missing parameter");
 			Boolean active = new Boolean(activeString);