+++ cmn-data-svc-server/src/test/java/org/acumos/cds/test/CdsControllerTest.java
@@ -86,6 +86,7 @@ import org.acumos.cds.domain.MLPUserNotification;
 import org.acumos.cds.domain.MLPValidationSequence;
 import org.acumos.cds.domain.MLPValidationStatus;
 import org.acumos.cds.domain.MLPValidationType;
+import org.acumos.cds.transport.AuthorTransport;
 import org.acumos.cds.transport.RestPageRequest;
 import org.acumos.cds.transport.RestPageResponse;
 import org.acumos.cds.transport.SuccessTransport;
@@ -164,8 +165,7 @@ public class CdsControllerTest {
 
 			MLPSolution cs = new MLPSolution();
 			cs.setName("solution name");
-			cs.setOwnerId(cu.getUserId());
-			cs.setProvider("Big Data Org");
+			cs.setUserId(cu.getUserId());
 			cs.setModelTypeCode("CL");
 			cs.setToolkitTypeCode("CP");
 			cs.setActive(true);
@@ -181,6 +181,7 @@ public class CdsControllerTest {
 			MLPSolutionRevision cr = new MLPSolutionRevision(cs.getSolutionId(), "1.0R", cu.getUserId(),
 					AccessTypeCode.PB.name(), ValidationStatusCode.IP.name());
 			cr.setDescription("Some description");
+			cr.setPublisher("Big Data Org");
 			cr = client.createSolutionRevision(cr);
 			logger.info("Created solution revision {}", cr);
 
@@ -479,7 +480,7 @@ public class CdsControllerTest {
 			ca.setVersion(version);
 			ca.setName("artifact name");
 			ca.setUri("http://nexus/artifact");
-			ca.setOwnerId(cu.getUserId());
+			ca.setUserId(cu.getUserId());
 			ca.setArtifactTypeCode(ArtifactTypeCode.DI.toString());
 			ca.setSize(1);
 			ca = client.createArtifact(ca);
@@ -503,7 +504,7 @@ public class CdsControllerTest {
 			ca2.setName("replicated artifact ");
 			ca2.setUri("http://other.foo");
 			ca2.setArtifactTypeCode(ArtifactTypeCode.CD.toString());
-			ca2.setOwnerId(cu.getUserId());
+			ca2.setUserId(cu.getUserId());
 			ca2.setSize(456);
 			ca2 = client.createArtifact(ca2);
 			Assert.assertEquals(artId, ca2.getArtifactId());
@@ -542,7 +543,7 @@ public class CdsControllerTest {
 			Assert.assertTrue(tags.getNumberOfElements() > 0);
 
 			MLPSolution cs = new MLPSolution("solution name", cu.getUserId(), true);
-			cs.setProvider("Tagged solution org");
+			cs.setDescription("Tagged solution");
 			cs.setModelTypeCode(ModelTypeCode.CL.name());
 			cs.setToolkitTypeCode(ToolkitTypeCode.CP.name());
 			cs.getTags().add(tag1);
@@ -553,7 +554,7 @@ public class CdsControllerTest {
 
 			// no tags
 			MLPSolution csOrg = new MLPSolution("solution organization", cu.getUserId(), true);
-			csOrg.setProvider("Untagged solution org");
+			csOrg.setDescription("Untagged solution org");
 			csOrg.setModelTypeCode(ModelTypeCode.DS.name());
 			csOrg.setToolkitTypeCode(ToolkitTypeCode.SK.name());
 			csOrg = client.createSolution(csOrg);
@@ -562,8 +563,8 @@ public class CdsControllerTest {
 
 			MLPSolution inactive = new MLPSolution();
 			inactive.setName("inactive solution name");
-			inactive.setOwnerId(cu.getUserId());
-			inactive.setProvider("Inactive Data Org");
+			inactive.setUserId(cu.getUserId());
+			inactive.setDescription("Inactive Data Org");
 			inactive.setModelTypeCode(ModelTypeCode.DS.name());
 			inactive.setToolkitTypeCode(ToolkitTypeCode.SK.name());
 			inactive.setActive(false);
@@ -631,18 +632,21 @@ public class CdsControllerTest {
 
 			MLPSolutionRevision cr = new MLPSolutionRevision(cs.getSolutionId(), "1.0R", cu.getUserId(), //
 					AccessTypeCode.PR.name(), ValidationStatusCode.NV.name());
+			cr.setAuthors(new AuthorTransport[] { new AuthorTransport("my name", "http://github") });
+			cr.setPublisher("publisher 1");
 			cr = client.createSolutionRevision(cr);
 			Assert.assertNotNull(cr.getRevisionId());
 			logger.info("Created solution revision {}", cr.getRevisionId());
+			cr.setDescription("Some description");
+			client.updateSolutionRevision(cr);
 
 			MLPSolutionRevision crOrg = new MLPSolutionRevision(csOrg.getSolutionId(), "1.0R", cu.getUserId(), //
 					AccessTypeCode.PR.name(), ValidationStatusCode.NV.name());
+			crOrg.setAuthors(new AuthorTransport[] { new AuthorTransport("your name", "email") });
+			crOrg.setPublisher("publisher 2");
 			crOrg = client.createSolutionRevision(crOrg);
 			Assert.assertNotNull(crOrg.getRevisionId());
 			logger.info("Created solution revision {}", cr.getRevisionId());
-
-			cr.setDescription("Some description");
-			client.updateSolutionRevision(cr);
 
 			logger.info("Adding artifact to revision 1");
 			client.addSolutionRevisionArtifact(cs.getSolutionId(), cr.getRevisionId(), ca.getArtifactId());
@@ -679,23 +683,23 @@ public class CdsControllerTest {
 			// Portal dynamic search
 			logger.info("Querying for any solutions via flexible i/f");
 			RestPageResponse<MLPSolution> portalAnyMatches = client.findPortalSolutions(null, null, true, null, null,
-					null, null, null, new RestPageRequest(0, 5));
+					null, null, null, null, null, new RestPageRequest(0, 5));
 			Assert.assertTrue(portalAnyMatches != null && portalAnyMatches.getNumberOfElements() > 1);
 
 			logger.info("Querying for valid tag on solutions via flexible i/f");
 			String[] searchTags = new String[] { tagName1 };
 			RestPageResponse<MLPSolution> portalTagMatches = client.findPortalSolutions(null, null, true, null, null,
-					null, null, searchTags, new RestPageRequest(0, 5));
+					null, null, searchTags, null, null, new RestPageRequest(0, 5));
 			Assert.assertTrue(portalTagMatches != null && portalTagMatches.getNumberOfElements() > 0);
 
 			logger.info("Querying for bogus tag on solutions via flexible i/f");
 			String[] bogusTags = new String[] { "bogus" };
 			RestPageResponse<MLPSolution> portalTagNoMatches = client.findPortalSolutions(null, null, true, null, null,
-					null, null, bogusTags, new RestPageRequest(0, 5));
+					null, null, bogusTags, null, null, new RestPageRequest(0, 5));
 			Assert.assertTrue(portalTagNoMatches != null && portalTagNoMatches.getNumberOfElements() == 0);
 
 			logger.info("Querying for user solutions via flexible i/f");
-			RestPageResponse<MLPSolution> userSols = client.findUserSolutions(null, null, true, 
+			RestPageResponse<MLPSolution> userSols = client.findUserSolutions(null, null, true,
 					inactiveUser.getUserId(), null, null, null, null, new RestPageRequest(0, 5));
 			Assert.assertTrue(userSols != null && userSols.getNumberOfElements() > 0);
 
@@ -705,11 +709,14 @@ public class CdsControllerTest {
 			String[] accessTypeCodes = { AccessTypeCode.PR.name() };
 			String[] modelTypeCodes = null;
 			String[] valStatusCodes = { ValidationStatusCode.NV.name() };
+			String[] authKw = { "github" };
+			String[] pubKw = { "publisher" };
 			searchTags = null;
 
 			// find active solutions
 			RestPageResponse<MLPSolution> portalActiveMatches = client.findPortalSolutions(nameKw, descKw, true, owners,
-					accessTypeCodes, modelTypeCodes, valStatusCodes, searchTags, new RestPageRequest(0, 5));
+					accessTypeCodes, modelTypeCodes, valStatusCodes, searchTags, authKw, pubKw,
+					new RestPageRequest(0, 5));
 			Assert.assertTrue(portalActiveMatches != null && portalActiveMatches.getNumberOfElements() > 0);
 
 			// Requires revisions and artifacts!
@@ -1427,8 +1434,7 @@ public class CdsControllerTest {
 				new RestPageRequest(0, 1));
 		Assert.assertTrue(threadComments != null && threadComments.hasContent());
 
-		long commentCountById = client.getSolutionRevisionCommentCount(cs.getSolutionId(),
-				cr.getRevisionId());
+		long commentCountById = client.getSolutionRevisionCommentCount(cs.getSolutionId(), cr.getRevisionId());
 		Assert.assertTrue(commentCountById > 0);
 
 		RestPageResponse<MLPComment> commentsById = client.getSolutionRevisionComments(cs.getSolutionId(),
@@ -2296,14 +2302,14 @@ public class CdsControllerTest {
 			logger.info("Create solution failed on dupe as expected: {}", ex.getResponseBodyAsString());
 		}
 		try {
-			cs.setOwnerId(s64);
+			cs.setUserId(s64);
 			client.updateSolution(cs);
 			throw new Exception("Unexpected success");
 		} catch (HttpStatusCodeException ex) {
 			logger.info("Update solution failed on constraints as expected: {}", ex.getResponseBodyAsString());
 		}
 		// restore valid value
-		cs.setOwnerId(cu.getUserId());
+		cs.setUserId(cu.getUserId());
 
 		try {
 			Map<String, Object> queryParameters = new HashMap<>();
@@ -2402,7 +2408,7 @@ public class CdsControllerTest {
 			logger.info("Update solution revision failed on empty as expected: {}", ex.getResponseBodyAsString());
 		}
 		try {
-			MLPSolutionRevision r = new MLPSolutionRevision(cs.getSolutionId(), "version", "ownerId",
+			MLPSolutionRevision r = new MLPSolutionRevision(cs.getSolutionId(), "version", "userId",
 					AccessTypeCode.PB.name(), ValidationStatusCode.NV.name());
 			r.setRevisionId("bogus");
 			client.updateSolutionRevision(r);
@@ -2538,14 +2544,14 @@ public class CdsControllerTest {
 
 		client.addSolutionRevisionArtifact(cs.getSolutionId(), csr.getRevisionId(), ca.getArtifactId());
 		try {
-			ca.setOwnerId(s64);
+			ca.setUserId(s64);
 			client.updateArtifact(ca);
 			throw new Exception("Unexpected success");
 		} catch (HttpStatusCodeException ex) {
 			logger.info("Update artifact failed on constraints as expected: {}", ex.getResponseBodyAsString());
 		}
 		// Restore valid value
-		ca.setOwnerId(cu.getUserId());
+		ca.setUserId(cu.getUserId());
 
 		try {
 			client.incrementSolutionViewCount("bogus");
@@ -2632,7 +2638,8 @@ public class CdsControllerTest {
 		}
 		try {
 			String[] searchTags = new String[] { "%" };
-			client.findPortalSolutions(null, null, true, null, null, null, null, searchTags, new RestPageRequest(0, 1));
+			client.findPortalSolutions(null, null, true, null, null, null, null, searchTags, null, null,
+					new RestPageRequest(0, 1));
 			// I have not been able to make findPortalSolutions fail.
 			// all arguments are optional; there is no illegal value; etc.
 			// so can't throw new Exception("Unexpected success") here.